refactor: Asset 파일명 일관성 확보 및 ProjectCard UI 개선
