+++ src/shared/assets/CloseIcon.tsx
