+++ src/shared/assets/index.ts
@@ -1,5 +1,5 @@
 export { default as Logo } from './Logo';
 export { default as PersonIcon } from './PersonIcon';
 export { default as ArrowIcon } from './ArrowIcon';
-export { default as Like } from './Like';
-export { default as Close } from './Close';
+export { default as LikeIcon } from './LikeIcon';
+export { default as CloseIcon } from './CloseIcon';
